requirements.txt
backports.zoneinfo==0.2.1
click==8.1.3
Flask==2.1.2
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg==3.0.15
psycopg2==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
zipp==3.8.1
